Support multiple booleans in TerminalBlock
Though it can not occur in the current pattern, it would be possible
that more than just one boolean variable is used in the TerminalBlock.
If that were to happen, it would be possible for the getJoinedVariable()
method to pick the wrong variable as the check variablee, as it picked a
random boolean variable used in the TerminalBlock, assuming there was
just one. To resolve this, the check variable is now derived from the
pattern and checked to be a valid variable for streaming instead.