The check var does not have to be streamable
The check variable does not have to be streamable, though it will
normally always meet the conditions required to be be a stream variable.
The check variable is already checked for all needed conditions in the
code: it must be a local variable that is not final and of a boolean
type, initialized with true or false.